double --> float, failing inverse test

+++ libololearning/include/DenseMatrix.h
@@ -4,29 +4,29 @@
 class DenseVector {
 public:
     DenseVector(size_t size);
-    DenseVector(double* data, size_t size);
+    DenseVector(float* data, size_t size);
     ~DenseVector();
 
-    double get(size_t i);
-    void set(size_t i, double val);
+    float get(size_t i);
+    void set(size_t i, float val);
     void swap(size_t i, size_t j);
 
-    double norm2();
+    float norm2();
 
-    double distance2(DenseVector other);
-    double dot(DenseVector other);
+    float distance2(DenseVector other);
+    float dot(DenseVector other);
 
     DenseVector subtract(DenseVector other, bool inplace);
 
     void printVector();
 
     size_t size();
-    double* getData();
+    float* getData();
 
     DenseVector copy();
 private:
     size_t _size;
-    double* _data;
+    float* _data;
 };
 
 class DenseMatrix;
@@ -42,22 +42,22 @@ class DenseMatrix {
 public:
 
     DenseMatrix(size_t nrow, size_t ncol);
-    DenseMatrix(double* data, size_t nrow, size_t ncol);
+    DenseMatrix(float* data, size_t nrow, size_t ncol);
     ~DenseMatrix();
 
     static DenseMatrix eye(size_t n);
     static DenseMatrix ones(size_t nrow, size_t ncol);
     
-    double get(size_t row, size_t col);
-    void set(size_t row, size_t col, double val);
+    float get(size_t row, size_t col);
+    void set(size_t row, size_t col, float val);
 
-    double norm2();
+    float norm2();
 
     DenseVector getColumn(size_t col);
     DenseVector getRow(size_t row);
     void swapRows(size_t i, size_t j);
 
-    double distance2(DenseMatrix other);
+    float distance2(DenseMatrix other);
     DenseMatrix subtract(DenseMatrix other, bool inplace);
 
     DenseMatrix transpose(); 
@@ -82,8 +82,8 @@ public:
     void printMatrix();
 
 private:
-    void init(double* data, size_t nrow, size_t ncol);
+    void init(float* data, size_t nrow, size_t ncol);
     size_t _nrow;
     size_t _ncol;
-    double* _data;
+    float* _data;
 };
